Bug fix resolves #130 (#133)
* Start device monitoring on initialization intead of starting of application to allow for GUI loop to start

* Allow running start redis server multiple times.

* Change logging to debug.

* - publish message function takes in message options parameter
- This allows specifying things like max retires

* Fixup: Need to initialize message_kwargs as empty dicg as well.

* Unwantd import removed

* - Change run files to import files in a clearer manner, and use context managers in one place.
- Also dramatiq_workers renamed to dramatiq_workers_context to clarify it as a context manager, not a function to initialize the workers directly.

* # Bug Fix:

- Resolves #130
- Need to move the chip insert/not status from the service function, to the listener routine. Before this fix, the chip insert message publishing was never reached since the listener actor filtered these out as well but did not publish the chip insertion message.

* fixup! - Change run files to import files in a clearer manner, and use context managers in one place. - Also dramatiq_workers renamed to dramatiq_workers_context to clarify it as a context manager, not a function to initialize the workers directly.

+++ microdrop_utils/broker_server_helpers.py
@@ -22,16 +22,20 @@ def is_redis_running():
 
 def start_redis_server(retries=5, wait=3):
     """Start the Redis server."""
-    process = subprocess.Popen(["redis-server", f"{os.path.dirname(__file__)}{os.sep}redis.conf"])
-    for _ in range(retries):  # Retry up to 5 times
-        if is_redis_running():
-            print("Redis server is running.")
-            return process
-        else:
-            print("Waiting for Redis server to start...")
-            time.sleep(wait)
-    print("Failed to start Redis server.")
-    process.terminate()
+    if not is_redis_running():
+        process = subprocess.Popen(["redis-server", f"{os.path.dirname(__file__)}{os.sep}redis.conf"])
+        for _ in range(retries):  # Retry up to 5 times
+            if is_redis_running():
+                print("Redis server is running.")
+                return process
+            else:
+                print("Waiting for Redis server to start...")
+                time.sleep(wait)
+        print("Failed to start Redis server.")
+        process.terminate()
+
+    else:
+        print("Redis server is already running.")
     return None
 
 
@@ -82,7 +86,7 @@ def redis_server_context():
 
 
 @contextmanager
-def dramatiq_workers(**kwargs):
+def dramatiq_workers_context(**kwargs):
     """
     Context manager for apps that make use of dramatiq. They need the workers to exist.
     """
@@ -126,6 +130,5 @@ if __name__ == "__main__":
                 exit(0)
 
 
-    with redis_server_context():
-        with dramatiq_workers():
+    with redis_server_context(), dramatiq_workers_context():
             example_app_routine()
